Meeting 09.11.

+++ web/src/main/java/hu/neuron/training/zooapp/web/servlet/DeleteZooServlet.java
@@ -30,8 +30,9 @@ public class DeleteZooServlet extends HttpServlet {
         ZooDao zooDao = ac.getBean(ZooDao.class);
         zooDao.connect();
 
-
-        zooDao.delete(zooDao.findById(id));
+        if (!zooDao.findById(id).equals(null)){
+            zooDao.delete(zooDao.findById(id));
+        }
 
         req.setAttribute("zoos", zooDao.getAll());
         manager.closeConnection();
